requirements.txt
python-docx
PySide6
